poling stations in RaspberryShake seismic network through web app endpoint
This Python script sets up a UDP server to receive messages, particularly those starting with "ALARM." Upon receiving such messages, it queries seismic data from a specified endpoint for a primary station and a list of secondary stations provided as command-line arguments. It then compares the seismic data from the primary station with the averages of the corresponding data from the secondary stations. If the differences in acceleration, velocity, and displacement are within 5% of the primary station's values, the script forwards the original message to another port specified as FORWARD_PORT. Otherwise, it logs that the secondary station data is not sufficiently similar to the primary station data. The script continuously listens for incoming packets while running.